+++ resources/ReleaseNotes.md
@@ -3,4 +3,4 @@
 - Rename `mag.IMAPTestingAnalysis` to `mag.IMAPAnalysis`
 - Fix issue with mode being mistakenly confused for "Ramp" when first range is `NaN`
 - Fix issue in `mag.graphics.sftPlots` with cropped analysis not defined when `Filter` is empty
-- Add system tests for `mag.IMAPAnalysis`
+- Add system and unit tests for `mag.IMAPAnalysis`